Move load autosave related functions to their own file
I think it makes more sense this way
