+++ addon/utils/save.py
@@ -65,37 +65,3 @@ def powersave():
         return ({'INFO'}, f'PowerSaved "{path.name}"', {'FINISHED'})
 
     return ({'WARNING'}, 'No PowerSave folder set', {'CANCELLED'})
-
-
-def should_load_autosave():
-    current_path = utils.files.as_path(bpy.data.filepath)
-    autosave_path = utils.files.with_autosave(current_path)
-
-    current_exists = current_path.is_file()
-    autosave_exists = autosave_path.is_file()
-    is_autosave = current_path.stem.endswith('_autosave')
-
-    if current_exists and autosave_exists and not is_autosave:
-        current_time = current_path.stat().st_mtime_ns
-        autosave_time = autosave_path.stat().st_mtime_ns
-        return autosave_time > current_time
-
-
-def load_autosave():
-    current_path = utils.files.as_path(bpy.data.filepath)
-    autosave_path = utils.files.with_autosave(current_path)
-
-    try:
-        bpy.ops.wm.open_mainfile(filepath=str(autosave_path))
-    except:
-        return ({'ERROR'}, f'Failed to load "{autosave_path.name}"', {'CANCELLED'})
-
-    new_path = increment_until_unique(current_path)
-
-    try:
-        bpy.ops.wm.save_as_mainfile(filepath=str(new_path))
-        utils.files.add_to_recent_files()
-    except:
-        return ({'ERROR'}, f'Failed to save "{new_path.name}"', {'CANCELLED'})
-
-    return ({'INFO'}, f'Loaded "{autosave_path.name}" as "{new_path.name}"', {'FINISHED'})
